hdctools: servod: Alias controls for uarts supported on servo V2.
Servo V2 has support for 4 uarts:
  1. Legacy uart from Servo V1 (40pin)
  2. CPU console uart from Servo V2 (42pin)
  3. EC console uart from Servo V2 (42pin)
  4. Atmega USB HID keyboard emulator (internal)

The legacy uart provided uart console access typically to the main
CPU's console.

In practice this was supported on ARM platforms.  However, X86
platforms were complicated by the console coming across LPC to either
a SuperIO or EC.  In the case of EC it required a 'debug' version of
the EC firmware.  Support on those platforms varies.

The new CPU & EC uarts provide console access for the V2 Debug
header requirements and current platforms with the 42pin header.

This CL provides control aliases for clarifying that:
  1. uart3_XXX -> legacy_uart_XXX
  2. uart2_XXX -> cpu_uart_XXX
  3. uart1_XXX -> ec_uart_XXX

BUG=chrome-os-partner:8256
TEST=manual,
connect servo V2 to DUT of choice and test uart.
For example,

  sudo servod -c servo_v2.xml
  dut-control legacy_debug_uart_en:on
  screen `dut-control legacy_debug_uart_pty | cut -d: -f2`
  <should see console output>

Change-Id: I05f4e5d8c49fe638835fa11b15e47f5cc3ea2210
Reviewed-on: https://gerrit.chromium.org/gerrit/17037
Commit-Ready: Todd Broch <tbroch@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>