hdctools: build: specify a soname for Linux shared libs
If a library is meant to be linked against, it should declare a DT_SONAME
ELF tag so that other tools can easily locate it even if the filename is
changed to something else.

Gentoo installs also check for this, so let's fix that QA warning.

BUG=None
TEST=`emerge hdctools` did not so a missing SONAME warning
TEST=`readelf -d /usr/lib64/libftdii2c.so` showed SONAME=libftdii2c.so

Change-Id: Ic4d1f072227af9c7d9a4f2b6e26ea5c365f6d520
Reviewed-on: https://gerrit.chromium.org/gerrit/18285
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>