hdctools: ftdiuart: quiet errno messages from EAGAIN/EWOULDBLOCK.
On occasion, the following message:
  'writing ftdi data to pty: Resource temporarily unavailable'

is seen when running servod.  It belongs to the polling loop of the
ftdiuart.c and is due to errno EAGAIN || EWOULDBLOCK being signalled.
They should be harmless and appear so as the console input and output
continues to function.

BUG=none
TEST=manual, start servod, connect to uart, remains functional and I
no longer see the message when leaving attached overnight.

Change-Id: I079e2011eee53ea17e300f2906d5485742d03c7f
Reviewed-on: https://gerrit.chromium.org/gerrit/22181
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Mark Hayter <mdhayter@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>