hdctools: ftdiuart: bugfix for USB busy.
This bug has been waiting for me.  I believe it surfaced due to a
number of factors:
1. More people using multiple servo's
2. Servo V2 providing access to 3 uarts (boat loads of ftdi_read_data)
calls.

Fix is simply to recognize that ftdi_read_data may fail due to EBUSY
errno which should not be fatal.  The only other place that its called
is in ftdii2c which will also retry the entire i2c transaction in that
case.

BUG=chrome-os-partner:10776
TEST=launch multiple servod from one host connected via the same USB
hub.

Connect to uart of first servod and it still functions as expected.

Change-Id: I1d3db501b83bcdab0bd9b7e87a36bab15babb876
Reviewed-on: https://gerrit.chromium.org/gerrit/25994
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Ready: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>