hdctools: servod: Add driver for Parrot EC for rec_mode switch
There is a hardware issue on Parrot that it can't poll the servo rec_mode
line. So Parrot EC has a work-around to accept an additional EC commands
for setting and getting the recovery mode state, like sending the following
commands to EC:
  fbf5=5a\r
  fbf6=a5\r

The next boot will trigger recovery mode. This change adds this support on
servod.

BUG=chrome-os-partner:11585
TEST=manual
$ sudo servod -b parrot -c servoflex_v2_r0_p50.xml
$ dut-control rec_mode
off
$ dut-control rec_mode:on
$ dut-control rec_mode
on
<< Reboot and enter recovery mode >>

Change-Id: I330bbf72f19acd3af9068d296050517c6824efeb
Reviewed-on: https://gerrit.chromium.org/gerrit/30987
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>