hdctools: servod: Change interface 3 default to GPIO for servo V1.
The 3rd interface on servo was planned to be shared between UART &
GPIOs.  That sharing was difficult to provide so the plan was to
switch the interface back and forth to provide mutually exclusive
access to either the GPIO or the UART.

Once servo V2 became available, the shared UART+GPIO functionality was
no longer necesary.  Furthermore all consumers wishing to have the
UART functionality have converted to using servo V2.

The remaining consumers of servo V1 are using the GPIO functionality
of interface 3.  This CL makes that the default to simplify having to
invoke servod differently to change it.

BUG=chromium-os:34834
TEST=manual
1. launch servod
2. dut-control | egrep "uart|dut_flag"

     cpu_uart_baudrate:ERR
     dut_flaga:off
     dut_flagb:off
     cpu_uart_pty:ERR
     cpu_uart_bits:ERR
     cpu_uart_sbits:ERR
     cpu_uart_parity:ERR

Note, ERR are expected as the UART interface is no longer the default.

Change-Id: I14b2e9906ac18bd179d1459637e2a23588e362e5
Reviewed-on: https://gerrit.chromium.org/gerrit/34192
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Mark Hayter <mdhayter@chromium.org>
Commit-Ready: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>