hdctools: ec: Change the return types of command-issuing methods to literal structs
We are going to export the command-issuing methods as getters/setters. The
original return type re.MatchObject is not reabable. Should make it to a
literal structure.

This change is to change the return type of these command-issuing methods to a
list of tuples, each of which contains the entire matched string and all the
subgroups of the match.

BUG=chromium-os:34787
TEST=run_remote_tests.sh --remote $DUT suite:faft_ec

Change-Id: I52bc70f601eb482008f0ce6ba6081152ac032163
Reviewed-on: https://gerrit.chromium.org/gerrit/34607
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Ready: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>