hdctools: select the right flex configuration for Spring
The Spring board is using a 50-pin flex cable with a special PCB adapter
for the connector.
As the flex configuration is somewhat difficult to remember, allow to
pass the board name to servod to select the right configuration.

BUG=None
TEST=sudo emerge hdctools
sudo servod -b spring
see EC serial flashing and SPI flashing are working

Change-Id: If82686d5473ffb13de46d1834c1ca683050ffe8a
Reviewed-on: https://gerrit.chromium.org/gerrit/37851
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>