Allow partial control names for dut-control get requests
Sometimes the user does not remember the exact control name. Obtaining
the entire control list takes time and produces a lot of results. This
change modifies dut-control behavior: when an unknown control's value
is requested, a list of similarly named controls (those where the
requested control is a substring) is printed.

BRANCH=none
BUG=None
TEST=manual
	$ dut-control xyz
	No control named "xyz"

	$ dut-control ec
	No control named "ec"
	Consider ec_spi ec_uart_baudrate ec_uart_bits ec_uart_cmd ec_uart_en ec_uart_parity ec_uart_pty ec_uart_regexp ec_uart_sbits ec_uart_timeout rec_mode sd_detect
	$ dut-control ec_spi
	ec_spi:off
	$ dut-control ec_spi ec_uart_baudrate ec_uart_bits
	ec_spi:off
	ec_uart_baudrate:115200
	ec_uart_bits:eight

Change-Id: I455353c34f82aa0fab5dd0df38add74436db3bb7
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41424
Reviewed-by: Vic Yang <victoryang@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>