Increase default UART timeout
When controlling a servo through a Beaglebone board, the delays caused
by the slow host add to the delays of the DUT, which causes tests
expecting the target reply in a fixed time to fail.

In most cases there is no point in limiting the time below few
seconds, as the code waits for the DUT issuing some text on the
console and the wait terminates as soon as the text is seen. The
timeout in this case is juts a fallback to avoid the test getting
stuck forever.

Tests requiring more fine grained timeout can always use the
`ec_uart_timeout' control to set the required value.

This change sets a longer default timeout and makes sure that the same
value is used in all default cases.

BUG=chromium-os:22475
TEST=manual

  . ran the modified autotest (tweaked it to configure proper
    timeouts) with and without a Beaglebone in the middle, both
    succeeded:

    $ run_remote_tests.sh --board=link --remote=172.22.50.140  firmware_ECBootTime \
        --args 'board=link servo_host=172.22.50.141 servo_port=9999'
    $ run_remote_tests.sh --board=link --remote=192.168.1.4 --allow_offline_remote \
       firmware_ECBootTime --servo --args 'board=link'

Change-Id: I26cf5b382ba7b2f94da45646acfbd2188a135660
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41568
Reviewed-by: Todd Broch <tbroch@chromium.org>