Do not reorder arguments when values are assigned
The earlier change (https://gerrit.chromium.org/gerrit/41409) tried to
be too fancy when rearranging arguments of dut-control. In case when
dut-control is invoked to set some values, the order of controls
becomes important.

Do not reorder controls if any of the arguments call for setting a
control value.

BUG=chromium-os:38081
TEST=manual
  . tried running

    cros_bundle_firmware -b daisy -d exynos5250-snow -w usb

   it does not fail to detect the board anymore and succeeds
   reprogramming it.

 . checked that the order does not change if there is a set operation
   argument:

    $ dut-control tp34 tp27 tp16
    tp16:on
    tp27:on
    tp34:on
    $ dut-control tp34 tp27 tp16:on
    tp34:on
    tp27:on
    $

Change-Id: Id49d4db7be91afe00d3fb1d5bb9b529931e2a472
Reviewed-on: https://gerrit.chromium.org/gerrit/41778
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>