Do not prevent servod startup on initialization failures
Sometimes servod is initialized when the DUT is not connected, or is
not fully active, for instance the EC is not programmed. This causes
startup failures and shutting down servod.

We might want to come up with yet another attribute in the XML files,
marking objects init of which do not have to succeed, but as a quick
fix let's just not terminate servo startup on any initialization
failure, but for now we should ignore initialization failures

BUG=chrome-os-partner:15610
TEST=manual

  . corrupted a Daisy's ec image and started up a servod controlling
   its servo. Observed the following line in the servod log:

   2013-01-31 14:35:01,023 - Servod - ERROR - Problem initializing lid_open -> yes :: Timeout waiting for EC response.

   but servod did not fail to come up. Then reprogrammed the DUT with
   the correct EC firmware using the same servo session and observed
   the DUT to come up. Then restarted servod and observe it initialize
   without the error message.

Change-Id: I89cd16026d791a667128035a90c4d71835ddfba8
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/42453
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>