Add ability to capture uart channels
With this change servod provides a facility which allows to start and
stop capturing data generated by a UART channel on the DUT. Originally
only the CPU interface is enabled.

Two new controls are being added:
  cpu_uart_capture, on/off - to start/stop capturing CPU UART output
  cpu_uart_stream - a read only control, returns a string of CPU
                    channel output accumulated since the previous
                    invocation

When capture is enabled, a new thread is started. The thread connects
to the serial interface the CPU UART is mapped to, opens it for
non-blocking reads and starts polling the interface for data. All
received data is put in a buffer, up to 20,000 bytes.

The client is supposed to be reading the data from the buffer using
the cpu_uart_stream control. If the buffer overflows, the reading
thread add an overflow indication to the buffer and stops adding mode
data to it, dropping the uart channel output on the floor until the
buffer is flashed by the client. Access to the buffer is protected by
a mutex lock.

The reading thread terminates when capture is stopped or when the
parent thread exits (which happens when servod is terminated for
whatever reason).

BUG=chromium-os:15610
TEST=manual
    . run the following commands
      $ sudo servod -d --board=daisy  # (in a separate terminal)
      $ dut-control cpu_uart_capture:on
      $ dut-control cpu_uart_stream
      cpu_uart_stream:''
      $ dut-control cold_reset:on
      $ dut-control cold_reset:off
      $ dut-control cpu_uart_stream
      cpu_uart_stream:'\x00\x00\n\r\n\rU-Boot SPL, board rev 0\n\r'
      $ dut-control cpu_uart_stream
      cpu_uart_stream:'Uncompressing Linux... done, booting the kernel.\r\n'
      $ dut-control cpu_uart_stream
      cpu_uart_stream:''
      $ dut-control cpu_uart_stream
      cpu_uart_stream:''
      $ dut-control cpu_uart_capture:off

     Observe servod log entries indicating capture thread start/stop

     Issue command

      $ dut-control cold_reset:on

     now terminate earlier started servod using ^C and observe that
     the capture thread is terminated.

Change-Id: Ie45a2493b8393f8155f56fa37dc88cd701400bff
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43083
Reviewed-by: Todd Broch <tbroch@chromium.org>