Add support for FT230X
The new FT230X (used in Toad devices) is using "CBUS" for its GPIOs
(same as FT232R), the code is updated to act accordingly.

The current libftdi (v1.0) doesn't have the support for FT230X yet.
(but it's already in upstream repository)
I workaround that issue by using the fact that with older libftdi
FT230X is detected as a FT232BM and we are not really using FT232BM in
any of our devices.

BUG=chrome-os-partner:17570
TEST=read GPIO values on a Toad cable (with the updated toad.xml
configuration done in the next CL)

Change-Id: I7be4e6b596101566d9fbe58b8344b6c3be0e91f6
Reviewed-on: https://gerrit.chromium.org/gerrit/47919
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>