Do not fail servod if the default port is not available
When starting servod, it is convenient to let the program find the
port to use if the default port is not available and the operator did
not pick a specific port.

This is especially handy when multiple servod instances need to be run.

Let the default port numbers be in the 9990..9999 range. For backwards
compatibility start with 9999 and check ports down from there (so that
port numbers stay 4 digits in size).

BUG=none
TEST=manual
  . started two servod instances without specifying the port in the
    command line, saw both start smoothly, on ports 9999 and 9998
    respectively.

  . start one servo instance (see it use port 9999 by default) and
    then try starting another one with '--port=9999' in the command
    line.

   Observe it terminate with:
   "2013-04-29 10:30:21,538 - servod - CRITICAL - Port 9999 is busy"

Change-Id: I976191bb7aeee05984ed6cb5585c92d8dbac18a6
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/49374
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>