Servo V3: Autodetect Servo V3 boards.
As part of running Servo V3 in the labs, we want servod to be able
to autodetect the V3 board and launch the correct config file for
this specific board.

This change adds the necessary values to the configuration dictionaries
to make this possible.

BUG=chromium:252379
TEST=ran sudo servod without supplying a config and it launched
successfully.

Change-Id: Ia86fbb30c6966d121db0941ec80f18bceb6d2b05
Reviewed-on: https://gerrit.chromium.org/gerrit/59454
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>