Do not proceed if board XML file is not found
With boards multiplying faster than XML files for them, it is better
to be safe than sorry, and stop if the requested board is not found.

A typical problem with continuing is when servod is invoked by
run_remote_tests.sh for pit. servod happily starts even though board
file for pit does not exist, leaving the operator puzzled and confused
(she should have started servod first passing --board=spring).

A board file for pit will be created shortly.

BUG=chrome-os-partner:21116,chromium:263104
TEST=manual
   $ sudo servod --board=peach_pix
   2013-07-24 10:35:06,309 - servod - INFO - Start
   getString params:
           index: 3
           len: 128
           langid: -1
   2013-07-24 10:35:06,312 - servod - INFO - Found servo, vid: 0x18d1 pid: 0x5002 sid: 868534-00014
   getString params:
           index: 3
           len: 128
           langid: -1
   2013-07-24 10:35:06,314 - servod - ERROR - No XML overlay for board peach_pix

Change-Id: I1d80b538870a1e0ad28b594dd111d6fe16e552ae
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62737