Servo V3: Add Init Values to GPIO's
Noticed that initial values of GPIO's could vary each beaglebone boot,
servod initialization, or by device and setup. To prevent this, we
now force initialize them on servod startup.

Also for V3, the map for usbkey is swapped so over-rode the mapping.

BUG=None
TEST=Using local AFE launched platform_InstallTestImage before the
patch which failed and then succeeded with this patch.

Change-Id: I9cd5618410ba8b2d0889261d2e878749c9256fcd
Reviewed-on: https://gerrit.chromium.org/gerrit/63280
Commit-Queue: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>