Servo V3: Pipe servo board version to servo_server
Refactored parts of servo initialization in order to allow the servo
server to know the servo board version.

This changes also allow xmlrpclib users the ability to determine the
servo board they are using. This is necessary for FAFT tests on V3 as
it has different configs to allow firmware flashing.

BUG=chromium:268534
TEST=local setup, using xmlrpclib, called server.get_version()

Change-Id: If4c77c7c654ef1b2ecb39e720b2b6d5a7cb9a41c
Reviewed-on: https://gerrit.chromium.org/gerrit/64835
Commit-Queue: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>