Support installing an upstart job when building for Chrome OS
This changes the build to include an option to install a servod
upstart job.  The job is tailored for execution on a Chrome OS host.
The option is used in the ebuild when the target is Chrome OS
(as opposed to targeting a build host inside the chroot).

BUG=chromium:272342
TEST=put proper content in /var/lib/servod/config, see servod start

Change-Id: I0ff48b7cc53b3aa2fa001e0ac6acac7933cf211f
Reviewed-on: https://chromium-review.googlesource.com/178919
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>