Add USB Type-C/PD controls for servod.
For some designs with USB Type-C/PD support, we have re-purposed spi1
(EC spi) to be an additional UART interface to communicate with the
Type-C/pD MCU's console.

This CL adds the necessary new controls to use that UART in addition
to renaming the associated buffers and setting there defaults to match
that of other UARTs (enabled by default).

Note in addition to sourcing this xml file when launching servod the
operator must also manually configure the interfaces.

A future CL _may_ change the SPI1 default to instead be a UART as
most/all? CrOS devices do NOT write there EC's via SPI but instead via
the UART (STM32xxx) or JTAG (TI Blizzard).

BUG=chrome-os-partner:28826
TEST=manual, launch servod & query usbpd_uart_XXX controls.  See
     expected defaults.

  sudo servod -s 641220-00045 -b samus -c usbpd.xml -i \
    "dummy ftdi_i2c dummy dummy dummy ftdi_uart ftdi_uart ftdi_uart" &

  for ctrl in `dut-control -i | grep usbpd_uart | cut -d ' ' -f 1` ; do
    dut-control $ctrl
  done

  usbpd_uart_baudrate:115200
  usbpd_uart_bits:eight
  usbpd_uart_buf_en:on
  usbpd_uart_buf_on_flex_en:on
  usbpd_uart_capture:off
  usbpd_uart_cmd:None
  Problem with ['usbpd_uart_multicmd'] :: Finding get function _Get_uart_multicmd
  usbpd_uart_parity:none
  usbpd_uart_pty:/dev/pts/10
  usbpd_uart_regexp:None
  usbpd_uart_sbits:one
  usbpd_uart_stream:''
  usbpd_uart_timeout:3
  usbpd_uart_vref:pp3300

Change-Id: I703964a6e8bbe3dab287be0f833c629bb9fe37dd
Reviewed-on: https://chromium-review.googlesource.com/199556
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>