Allowing progress if no value provide for arg usbkm232 to servod.
For keyboardless devices, they need a special keyboard handler which uses a
USB km232 cable to send keystrokes. The path to the usbkm232 device is provided
at startup time via a --usbkm232 command line argument.

For the board types that required this cable, we would raise an Exception if the
argument was not specified. However, in the lab, this cabling has not been
setup and so starting up of servod is blocked.

Replacing the exception with a warning message and returning the default
keyboard handler if the usbkm232 arg is not specified. Hopefully the only
effect is any tests that require using the keyboard control of the DUT will not
work correctly. These should only be the FAFT tests right now.

BUG=chromium:342875
TEST=Tested at desk, starting up servod against --board=pather, without
specifying the usbkm232 arg.

Change-Id: Iff11488b3841d4687af42f885180db10b31d2ffc
Reviewed-on: https://chromium-review.googlesource.com/201074
Reviewed-by: Vic Yang <victoryang@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Yusuf Mohsinally <mohsinally@chromium.org>