cros_ec_power: add shutdown_delay parameter
For some systems the time to shutdown can be significant, this allows
the specification of a shutdown_delay parameter which allows the
power_state:off setting to wait until the system is fully off.

BUG=chromium:378048
TEST=Manually change the delay value to 10 and run:
`dut-control power_state:off power_state:on`
Observe that dut-control waits for 10s before powering the DUT back on.
Also run with the default of 0 and observe no delay.

Change-Id: I487174d1ccf41604454b2ffd66e51135ff2651e8
Reviewed-on: https://chromium-review.googlesource.com/201994
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>