Rambi: add 11s shutdown delay value
The Rambi based systems take ~10s to fully shut down from servo, to
the point they can be powered back on with a servo power button press,
this adds a 11s delay to ensure that when `dut-control power_state:off`
is run the device is actually off when the command completes.

BUG=chromium:378048
TEST=`dut-control power_state:off power_state:on`
Observe DUT actually come back on.

Change-Id: Ie55e6d290bb1428308308f553bce20c7410ea008
Reviewed-on: https://chromium-review.googlesource.com/201995
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>