veyron: add ina and servo support
This adds support for servo for veyron and defines the inas on the proto 0
board. The values for the inas are with the rework that replaces the 200 mOhm
resisitor with a 10 mOhm resistor.

BUG=None
TEST=run 'cros_workon --host start hdctools'
'sudo emerge hdctools'
'sudo servod --board veyron -c veyron_p0.xml'
'dut-control | grep _mw'
Make sure there are reasonable values for vcc_* and vdd_* values

Change-Id: I88d2df9d3d54a0ba76dd102665dccc710a721c6f
Reviewed-on: https://chromium-review.googlesource.com/209904
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>