Clean up board overlay alias handling.
Previously, there was a 'dut.py' module that provided a mapping of
board name aliases to board overlay names.  The aliases provided
were without rhyme or reason, and the mechanism itself was
redundant, since any alias could also be provided with a named
overlay file that included the alias' target.

This change removes the `dut.py` module, and requires all aliases
be implemented with an xml overlay named for the alias.  This preserves
all names that were present in dut.py, and adds a few more useful
aliases.

BUG=None
TEST=Run `servod --board=...` for all the new/changed board names

Change-Id: I3a249983dffff459ecc9e70b99f64dd62f5cbcf5
Reviewed-on: https://chromium-review.googlesource.com/246742
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Shawn N <shawnn@chromium.org>