keyboard: Simplify the key-simulating flow on devices with Chrome EC
The original servo design to simulate a key press pulls the keyboard
matrix of the internal keyboard. It sets proper rows and columns,
like simulating Ctrl-D by setting:
  kbd_m1_a0:0
  kbd_m1_a1:0
  kbd_m2_a0:0
  kbd_m2_a1:1

On devices with Chrome EC, the flow is still compatible with the
original matrix design. We implement a driver to interpret the
kbd_mx_ax's and send proper EC commands.

The servo keyboard interface was abstracted that allows different
keyboard implementations. So the flow on Chrome EC devices can be
simplified, to skip the kbd_mx_ax approach and call EC commands
directly. One more benefit is that we can simulate all the key
combinations, like Ctrl-U, which is not supported in the matrix
approach.

This change also moves the duplicated methods from multiple
subclasses to the base class.

BUG=chrome-os-partner:39653
TEST=Deploy the modified code to servod. Run firmware_DevMode to exercise the
Ctrl-D and Enter path. Run firmware_DevBootUSB to exercise the Ctrl-U path.
Ran these tests on Jerry which uses Chrome EC to exercise ChromeECHandler.
Modified the code to force using MatrixKeyboardHandler to exercise it.

Change-Id: Ie338463d9ba45a314d4772efadf908dad184787c
Reviewed-on: https://chromium-review.googlesource.com/267813
Trybot-Ready: Tom Tam <waihong@google.com>
Tested-by: Tom Tam <waihong@google.com>
Reviewed-by: danny chan <dchan@chromium.org>
Commit-Queue: Tom Tam <waihong@google.com>