CFLAGS: Do not use Wswitch-enum in GCC flags
-Wswitch, which is still left enabled, warns when a switch statement
lacks a "case:" for one or more enumerations, but acceps a "default:"
label.
-Wswitch-enum, creates the same warning, even when a "default:" label
is given, however, "-Wswitch-default", warns when the "default:" label
is missing.

One of these three compiler switches is thus redundant. Considering
that "-Wswitch-enum" had to be selectively disabled in a few places,
I've chosen to kill this flag. This also allows us to remove the hacks
around a few switch() statements.

BUG=none
TEST=Succesfully built outside of chroot ($ make -j8)

Change-Id: I848f9ba7014a600b8691abf4ad3ecaa2d73ee7d8
Signed-off-by: Alexandru Gagniuc <alexandrux.gagniuc@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/298515
Commit-Ready: Stefan Reinauer <reinauer@chromium.org>
Tested-by: AlexandruX Gagniuc <alexandrux.gagniuc@intel.com>
Reviewed-by: Todd Broch <tbroch@chromium.org>