servo: Add control for EC-3PO interp dis/connect.
There's a script in the EC repo, flash_ec, which is used to program
the ECs.  For stm32 targets, we often program these ECs using the
UART.  The EC-3PO interpreter is attached to these UARTs and therefore
the flash_ec script will need to temporarily disconnect the
interpreter from the UART.  Otherwise, the interpreter will interfere
with the programming and cause it to always fail.

This commit adds a new control which will allow the interpreter to be
disconnected and reconnected to the EC UART by means of dut-control.

BUG=chromium:571170
BRANCH=None
TEST=Use modified flash_ec script to program both veyron_jerry and
samus_pd with no issues with servo v2.

CQ-DEPEND=CL:318900

Change-Id: I09107f8a254d9739394cee110d1ce47b3688eff9
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/321091
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>