Fix the power_state driver for veyron laptops.
Veyron-based laptop boards misbehave if powered off twice in a row,
with the result that pressing the power button doesn't work as
expected.  This fixes the _power_on() method for veyron power_state
drivers to use the EC 'power on' command, instead of the power
button.

BUG=chromium:574242
TEST=TBD

Change-Id: I31e9990664264b02f8da53be820b5fa9ef3dd018
Reviewed-on: https://chromium-review.googlesource.com/321807
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Gediminas Ramanauskas <gedis@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>