plankton: Fix issue with _get_pd_state() in plankton servod driver
The method _get_pd_state() depended on a particular output string
which has since been modifed in PD firmware. I changed the method
to work the same way as what's done in the PlanktonConsole class.
This version uses a regexp which works for both the old and new
version of PD firmware and then parses the string for the desired
values to fill the dictionary.

BRANCH=none
BUG=chrome-os-partner:49300
TEST=dut-control pd_state

Change-Id: Ia3353ae3219c0ba4fca80b0ea1dbd00f670c1ea7
Signed-off-by: Scott Collyer <scollyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/321389
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>